+++ plugins/org.eclipse.glsp.server/src/org/eclipse/glsp/server/gmodel/GModelReconnectEdgeOperationHandler.java
@@ -1,5 +1,5 @@
 /********************************************************************************
- * Copyright (c) 2019-2021 EclipseSource and others.
+ * Copyright (c) 2019-2022 EclipseSource and others.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0 which is available at
@@ -13,7 +13,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  ********************************************************************************/
-package org.eclipse.glsp.server.operations.gmodel;
+package org.eclipse.glsp.server.gmodel;
 
 import static org.eclipse.glsp.server.types.GLSPServerException.getOrThrow;
 import static org.eclipse.glsp.server.utils.GModelUtil.IS_CONNECTABLE;
@@ -27,7 +27,10 @@ import org.eclipse.glsp.server.operations.ReconnectEdgeOperation;
 
 import com.google.inject.Inject;
 
-public class ReconnectEdgeOperationHandler extends AbstractOperationHandler<ReconnectEdgeOperation> {
+/**
+ * Applies {@link ReconnectEdgeOperation} directly to the GModel.
+ */
+public class GModelReconnectEdgeOperationHandler extends AbstractOperationHandler<ReconnectEdgeOperation> {
 
    @Inject
    protected GModelState modelState;