+++ plugins/org.eclipse.glsp.server/src/org/eclipse/glsp/server/types/ShapeTypeHint.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2019-2021 EclipseSource and others.
+ * Copyright (c) 2019-2022 EclipseSource and others.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0 which is available at
@@ -17,6 +17,9 @@ package org.eclipse.glsp.server.types;
 
 import java.util.List;
 
+/**
+ * {@link ElementTypeHint Element type hints} for shapes (nodes).
+ */
 public class ShapeTypeHint extends ElementTypeHint {
 
    private boolean resizable;